Create a new ValidatorConfigCache class and pass metadata validators in the constructor.
That way, we can keep the isFresh() method and get away without deprecations.
